+++ src/main/java/me/sallim/api/domain/chat/dto/response/ChatRoomResponse.java
@@ -1,4 +1,4 @@
-package me.sallim.api.domain.chat.dto;
+package me.sallim.api.domain.chat.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor
-public class ChatMessageResponseDto {
+public class ChatRoomResponse {
     private Long id;
     private Long senderId;
     private Long receiverId;
